feat: rename library from cbytes to bytekit and update related references
