+++ tests/copy_test.c
@@ -1,4 +1,4 @@
-#include <cbytes.h>
+#include <bytekit.h>
 #include "utils/log.c"
 
 void copy_test(const char *data)
