+++ tests/index_of_test.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-#include <cbytes.h>
+#include <bytekit.h>
 #include "utils/log.c"
 
 void insert_lookup_string(char *chars);
