+++ tests/length_test.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-#include <cbytes.h>
+#include <bytekit.h>
 #include "utils/log.c"
 
 void length_test(const char *data)
