+++ tests/main.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-#include <cbytes.h>
+#include <bytekit.h>
 #include "copy_test.c"
 #include "length_test.c"
 #include "index_of_test.c"
