+++ tests/test_copy.c
@@ -1,4 +1,4 @@
-#include <cbytes.h>
+#include <bytekit.h>
 #include "utils/log.c"
 
 int test_copy() {
