+++ tests/test_equal.c
@@ -1,4 +1,4 @@
-#include <cbytes.h>
+#include <bytekit.h>
 #include "utils/log.c"
 
 int test_equal() {
